+++ backend/temporal.py
@@ -4,12 +4,12 @@ def extract_temporal_patterns(df):
     """
     Extracts citywide hourly, daily, station-wide violation counts and date ranges.
     """
-    hourly_city = df["hour"].value_counts().sort_index().to_dict()
+    hourly_city = df["hour"].astype(int).value_counts().sort_index().to_dict()
     daily_city = df["day_of_week"].value_counts().to_dict()
     station_counts = df["police_station"].value_counts().head(15).to_dict()
 
     return {
-        "hourly_city": {str(k): int(v) for k, v in hourly_city.items()},
+        "hourly_city": {str(int(k)): int(v) for k, v in hourly_city.items()},
         "daily_city": daily_city,
         "station_counts": station_counts,
         "total_violations": len(df),
